+++ src/parseDdt.h
@@ -61,10 +61,6 @@ static inline int add_deState(nodeInfo ni, char *name, char *v, int hasLhs, int 
   if (fromWhere == fromDDT && strncmp(v, "rx__sens_", 3) == 0) {
     tb.sensi++;
   }
-  if (rx_syntax_allow_dots == 0 && strstr(v, ".")) {
-    updateSyntaxCol();
-    trans_syntax_error_report_fn(NODOT);
-  }
   new_or_ith(v);
   if (((tb.ini[tb.ix] == 1 && tb.ini0[tb.ix] == 0) ||
        (tb.lh[tb.ix] == isLHS || tb.lh[tb.ix] == isLHSparam))){