+++ src/tran.c
@@ -61,8 +61,7 @@ int rx_syntax_error = 0, rx_suppress_syntax_info=0, rx_podo = 0, rx_syntax_requi
 extern D_ParserTables parser_tables_rxode2;
 
 unsigned int found_jac = 0, nmtime=0;
-int rx_syntax_allow_dots = 0,
-  rx_syntax_allow_ini = 1, 
+int rx_syntax_allow_ini = 1, 
   maxSumProdN = 0, SumProdLD = 0, good_jac=1, extraCmt=0;
 
 sbuf s_inits;
@@ -136,7 +135,6 @@ static inline int isCmtLhsStatement(nodeInfo ni, char *name, char *v);
 //static inline int add_deCmtProp(nodeInfo ni, char *name, char *v, int hasLhs, int fromWhere);
 static inline void add_de(nodeInfo ni, char *name, char *v, int hasLhs, int fromWhere);
 
-
 #include "parseFuns.h"
 #include "parseLogical.h"
 #include "parseIdentifier.h"
@@ -145,7 +143,6 @@ static inline void add_de(nodeInfo ni, char *name, char *v, int hasLhs, int from
 #include "parseDfdy.h"
 #include "parseCmtProperties.h"
 #include "parseDdt.h"
-
 
 static inline int parseNodePossiblySkipRecursion(nodeInfo ni, char *name, D_ParseNode *pn, D_ParseNode *xpn,
 						 int *i, int nch, int *depth) {
@@ -370,7 +367,6 @@ void reset() {
   rx_syntax_error = 0;
   rx_suppress_syntax_info=0;
   rx_podo = 0;
-  rx_syntax_allow_dots = 0;
   rx_syntax_allow_ini = 1;
 
   maxSumProdN = 0;
@@ -492,7 +488,6 @@ static inline int setupTrans(SEXP parse_file, SEXP prefix, SEXP model_md5, SEXP 
 
   int isStr =INTEGER(parseStr)[0];
   reset();
-  rx_syntax_allow_dots = R_get_option("rxode2.syntax.allow.dots",1);
   rx_suppress_syntax_info = R_get_option("rxode2.suppress.syntax.info",0);
   rx_syntax_allow_ini  = R_get_option("rxode2.syntax.allow.ini",1);
   rx_syntax_require_ode_first = R_get_option("rxode2.syntax.require.ode.first",1);